requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-slugify==8.0.4
gunicorn==23.0.0
PyMySQL==1.1.1
